MNMS-655 rename: swagger 파일 경로 변경
