+++ src/main/java/com/tekcit/festival/domain/user/controller/EmailController.java
@@ -1,13 +1,12 @@
 package com.tekcit.festival.domain.user.controller;
 
-import com.tekcit.festival.domain.user.dto.api.EmailApiSpecification;
+import com.tekcit.festival.domain.user.controller.api.EmailApiSpecification;
 import com.tekcit.festival.domain.user.dto.request.EmailSendDTO;
 import com.tekcit.festival.domain.user.dto.request.EmailVerifyDTO;
 import com.tekcit.festival.domain.user.dto.response.EmailResponseDTO;
 import com.tekcit.festival.domain.user.service.EmailService;
 import com.tekcit.festival.exception.global.SuccessResponse;
 import com.tekcit.festival.utils.ApiResponseUtil;
-import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
