+++ src/main/java/com/tekcit/festival/domain/user/controller/api/EmailApiSpecification.java
@@ -1,4 +1,4 @@
-package com.tekcit.festival.domain.user.dto.api;
+package com.tekcit.festival.domain.user.controller.api;
 
 import com.tekcit.festival.domain.user.dto.request.EmailSendDTO;
 import com.tekcit.festival.domain.user.dto.request.EmailVerifyDTO;
